refactor and patch job output checking
