+++ plugins/modules/zos_job_output.py
@@ -113,164 +113,9 @@ zos_job_output:
 
 import json
 from ansible.module_utils.basic import AnsibleModule
+from ansible_collections.ibm.ibm_zos_core.plugins.module_utils.job import job_output
 from os import chmod, path, remove
 from tempfile import NamedTemporaryFile
-
-
-def get_job_json(jobid, owner, jobname, ddname, module):
-    get_job_detail_json_rexx = """/* REXX */
-arg options
-parse var options param
-upper param
-parse var param 'JOBID=' jobid ' OWNER=' owner,
-  ' JOBNAME=' jobname ' DDNAME=' ddname
-
-rc=isfcalls('ON')
-
-jobid = strip(jobid,'L')
-if (jobid <> '') then do
-  ISFFILTER='JobID EQ '||jobid
-end
-owner = strip(owner,'L')
-if (owner <> '') then do
-  ISFOWNER=owner
-end
-jobname = strip(jobname,'L')
-if (jobname <> '') then do
-  ISFPREFIX=jobname
-end
-ddname = strip(ddname,'L')
-if (ddname == '?') then do
-  ddname = ''
-end
-
-Address SDSF "ISFEXEC ST (ALTERNATE DELAYED)"
-if rc<>0 then do
-  Say '{"jobs":[]}'
-  Exit 0
-end
-
-if isfrows == 0 then do
-  Say '{"jobs":[]}'
-end
-else do
-  Say '{"jobs":['
-  do ix=1 to isfrows
-    if ix<>1 then do
-      Say ','
-    end
-    Say '{'
-    Say '"'||'job_id'||'":"'||value('JOBID'||"."||ix)||'",'
-    Say '"'||'job_name'||'":"'||value('JNAME'||"."||ix)||'",'
-    Say '"'||'subsystem'||'":"'||value('ESYSID'||"."||ix)||'",'
-    Say '"'||'owner'||'":"'||value('OWNERID'||"."||ix)||'",'
-    Say '"'||'ret_code'||'":{"'||'msg'||'":"'||value('RETCODE'||"."||ix)||'"},'
-    Say '"'||'class'||'":"'||value('JCLASS'||"."||ix)||'",'
-    Say '"'||'content-type'||'":"'||value('JTYPE'||"."||ix)||'",'
-    Say '"'||'changed'||'":"'||'false'||'",'
-    Say '"'||'failed'||'":"'||'false'||'",'
-    Address SDSF "ISFACT ST TOKEN('"TOKEN.ix"') PARM(NP ?)",
-"("prefix JDS_
-    lrc=rc
-    if lrc<>0 | JDS_DDNAME.0 == 0 then do
-      Say '"ddnames":[]'
-    end
-    else do
-      Say '"ddnames":['
-      do jx=1 to JDS_DDNAME.0
-        if jx<>1 & ddname == '' then do
-          Say ','
-        end
-        if ddname == '' | ddname == value('JDS_DDNAME'||"."||jx) then do
-          Say '{'
-          Say '"'||'ddname'||'":"'||value('JDS_DDNAME'||"."||jx)||'",'
-          Say '"'||'record-count'||'":"'||value('JDS_RECCNT'||"."||jx)||'",'
-          Say '"'||'id'||'":"'||value('JDS_DSID'||"."||jx)||'",'
-          Say '"'||'stepname'||'":"'||value('JDS_STEPN'||"."||jx)||'",'
-          Say '"'||'procstep'||'":"'||value('JDS_PROCS'||"."||jx)||'",'
-          Say '"'||'byte-count'||'":"'||value('JDS_BYTECNT'||"."||jx)||'",'
-          Say '"'||'content'||'":['
-          Address SDSF "ISFBROWSE ST TOKEN('"TOKEN.ix"')"
-          do kx=1 to isfline.0
-            if kx<>1 then do
-              Say ','
-            end
-            Say '"'||escapeNewLine(escapeDoubleQuote(isfline.kx))||'"'
-          end
-          Say ']'
-          Say '}'
-        end
-      end
-      Say ']'
-    end
-    Say '}'
-  end
-  Say ']}'
-end
-
-rc=isfcalls('OFF')
-
-return 0
-
-escapeDoubleQuote: Procedure
-Parse Arg string
-out=''
-Do While Pos('"',string)<>0
-   Parse Var string prefix '"' string
-   out=out||prefix||'\\"'
-End
-Return out||string
-
-escapeNewLine: Procedure
-Parse Arg string
-Return translate(string, '4040'x, '1525'x)
-"""
-    try:
-        dirname, scriptname = _copy_temp_file(get_job_detail_json_rexx)
-        if jobid is None:
-            jobid = ''
-        if owner is None:
-            owner = ''
-        if jobname is None:
-            jobname = ''
-        if ddname is None or ddname == '?':
-            ddname = ''
-        jobid_param = 'jobid=' + jobid
-        owner_param = 'owner=' + owner
-        jobname_param = 'jobname=' + jobname
-        ddname_param = 'ddname=' + ddname
-        cmd = [dirname + '/' + scriptname, jobid_param, owner_param,
-               jobname_param, ddname_param]
-
-        rc, out, err = module.run_command(args=_list_to_string(cmd),
-                                          cwd=dirname,
-                                          use_unsafe_shell=True)
-    except Exception:
-        raise
-    finally:
-        remove(dirname + "/" + scriptname)
-    return rc, out, err
-
-
-def _list_to_string(s):
-    str1 = " "
-    return str1.join(s)
-
-
-def _copy_temp_file(content):
-    delete_on_close = False
-    try:
-        tmp_file = NamedTemporaryFile(delete=delete_on_close)
-        with open(tmp_file.name, 'w') as f:
-            f.write(content)
-        f.close()
-        chmod(tmp_file.name, 0o755)
-        dirname = path.dirname(tmp_file.name)
-        scriptname = path.basename(tmp_file.name)
-    except Exception:
-        remove(tmp_file)
-        raise
-    return dirname, scriptname
 
 
 def run_module():
@@ -283,29 +128,19 @@ def run_module():
 
     module = AnsibleModule(argument_spec=module_args, supports_check_mode=True)
 
-    job_id = module.params.get("job_id")
-    job_name = module.params.get("job_name")
-    owner = module.params.get("owner")
-    ddname = module.params.get("ddname")
+    job_id = module.params.get("job_id") or ""
+    job_name = module.params.get("job_name") or ""
+    owner = module.params.get("owner") or ""
+    ddname = module.params.get("ddname") or ""
 
-    if job_id is None and job_name is None and owner is None:
+    if not job_id and not job_name and not owner:
         module.fail_json(msg="Please provide a job_id or job_name or owner")
 
-    job_detail_json = {}
-    rc, out, err = get_job_json(job_id, owner,
-                                job_name, ddname, module)
-    if rc != 0:
-        module.fail_json(msg=err, rc=rc)
-    if not out:
-        module.fail_json(msg="No job output was found.")
-    else:
-        job_detail_json = json.loads(out, strict=False)
-
+    try:
+        job_detail_json = job_output(module, job_id, owner, job_name, ddname)
+    except Exception as e:
+        module.fail_json(msg=str(e))
     module.exit_json(zos_job_output=job_detail_json)
-
-
-class Error(Exception):
-    pass
 
 
 def main():
